requirements.txt
psycopg2-binary~=2.9.11
requests~=2.32.5
apscheduler~=3.11.1
fastapi~=0.123.4
uvicorn[standard]~=0.23.2
python-dotenv~=1.2.1
pandas~=2.3.3
boto3~=1.42.6
numpy~=2.3.5
torch~=2.9.1
joblib~=1.5.2